pb libft strdup vs strlen bizarre

+++ libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: thifranc <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/02/14 11:34:52 by thifranc          #+#    #+#             */
-/*   Updated: 2016/04/14 10:32:36 by thifranc         ###   ########.fr       */
+/*   Updated: 2016/04/14 15:37:40 by thifranc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *c)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	while (c && c[i])
